feat: added test

+++ db/migrate/20250731145640_remove_old_starting_invoice_subscription_index.rb
@@ -4,7 +4,9 @@ class RemoveOldStartingInvoiceSubscriptionIndex < ActiveRecord::Migration[8.0]
   disable_ddl_transaction!
 
   def up
-    remove_index :invoice_subscriptions, name: :index_unique_starting_subscription_invoice
+    if index_exists?(:invoice_subscriptions, nil, name: :index_unique_starting_subscription_invoice)
+      remove_index :invoice_subscriptions, name: :index_unique_starting_subscription_invoice
+    end
   end
 
   def down
@@ -13,6 +15,7 @@ class RemoveOldStartingInvoiceSubscriptionIndex < ActiveRecord::Migration[8.0]
       unique: true,
       name: :index_unique_starting_subscription_invoice,
       where: "invoicing_reason = 'subscription_starting'",
-      algorithm: :concurrently
+      algorithm: :concurrently,
+      if_not_exists: true
   end
 end
